+++ packages/elements/progress-tracker/examples/basic.js
@@ -1,6 +1,6 @@
 // @flow
 import React from 'react';
-import ProgressTracker from '../src/ProgressTracker';
+import { ProgressTracker } from '../src';
 
 const items = [
   {