+++ packages/fabric/editor-core/__tests__/editor/create-editor/create-plugins-list.ts
@@ -7,7 +7,7 @@ import {
   tablePlugin,
   mediaPlugin,
   helpDialogPlugin,
-  placeholderCursorPlugin,
+  fakeTextCursorPlugin,
   submitEditorPlugin,
 } from '../../../src/editor/plugins';
 
@@ -19,9 +19,9 @@ describe('createPluginsList', () => {
     expect(plugins).toContain(helpDialogPlugin);
   });
 
-  it('should add placeholderCursorPlugin if allowPlaceholderCursor is true', () => {
-    const plugins = createPluginsList({ allowPlaceholderCursor: true });
-    expect(plugins).toContain(placeholderCursorPlugin);
+  it('should add fakeTextCursorPlugin by default', () => {
+    const plugins = createPluginsList({});
+    expect(plugins).toContain(fakeTextCursorPlugin);
   });
 
   it('should add tablePlugin if allowTables is true', () => {