+++ packages/fabric/editor-core/src/editor/create-editor/create-plugins-list.ts
@@ -26,7 +26,7 @@ import {
   panelPlugin,
   macroPlugin,
   confluenceInlineComment,
-  placeholderCursorPlugin,
+  fakeTextCursorPlugin,
   extensionPlugin,
   rulePlugin,
   clearMarksOnChangeToEmptyDocumentPlugin,
@@ -142,10 +142,6 @@ export default function createPluginsList(props: EditorProps): EditorPlugin[] {
     plugins.push(confluenceInlineComment);
   }
 
-  if (props.allowPlaceholderCursor) {
-    plugins.push(placeholderCursorPlugin);
-  }
-
   if (props.allowDate) {
     plugins.push(datePlugin);
   }
@@ -158,6 +154,7 @@ export default function createPluginsList(props: EditorProps): EditorPlugin[] {
   );
 
   plugins.push(submitEditorPlugin);
+  plugins.push(fakeTextCursorPlugin);
 
   return plugins;
 }